+++ src/hapi/pipelines/app/pipelines.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from typing import Dict, Optional
 
+from hdx.api.configuration import Configuration
 from hdx.location.adminlevel import AdminLevel
 from hdx.scraper.runner import Runner
 from hdx.scraper.utilities.sources import Sources
@@ -9,12 +10,8 @@ from hdx.utilities.typehint import ListTuple
 from sqlalchemy.orm import Session
 
 from hapi.pipelines.database.admins import Admins
-from hapi.pipelines.database.age_range import AgeRange
 from hapi.pipelines.database.food_security import FoodSecurity
-from hapi.pipelines.database.gender import Gender
 from hapi.pipelines.database.humanitarian_needs import HumanitarianNeeds
-from hapi.pipelines.database.ipc_phase import IpcPhase
-from hapi.pipelines.database.ipc_type import IpcType
 from hapi.pipelines.database.locations import Locations
 from hapi.pipelines.database.metadata import Metadata
 from hapi.pipelines.database.national_risk import NationalRisk
@@ -22,15 +19,13 @@ from hapi.pipelines.database.operational_presence import OperationalPresence
 from hapi.pipelines.database.org import Org
 from hapi.pipelines.database.org_type import OrgType
 from hapi.pipelines.database.population import Population
-from hapi.pipelines.database.population_group import PopulationGroup
-from hapi.pipelines.database.population_status import PopulationStatus
 from hapi.pipelines.database.sector import Sector
 
 
 class Pipelines:
     def __init__(
         self,
-        configuration: Dict,
+        configuration: Configuration,
         session: Session,
         today: datetime,
         themes_to_run: Optional[Dict] = None,
@@ -44,13 +39,12 @@ class Pipelines:
         self.locations = Locations(
             configuration=configuration,
             session=session,
-            today=today,
             use_live=use_live,
         )
         countries = configuration["HAPI_countries"]
         libhxl_dataset = AdminLevel.get_libhxl_dataset().cache()
         self.admins = Admins(
-            configuration, session, self.locations, libhxl_dataset, today
+            configuration, session, self.locations, libhxl_dataset
         )
         self.adminone = AdminLevel(admin_level=1)
         self.admintwo = AdminLevel(admin_level=2)
@@ -60,22 +54,9 @@ class Pipelines:
         self.admintwo.load_pcode_formats()
         self.admintwo.set_parent_admins_from_adminlevels([self.adminone])
 
-        self.population_status = PopulationStatus(
-            session=session,
-            population_status_descriptions=configuration[
-                "population_status_descriptions"
-            ],
-        )
-        self.population_group = PopulationGroup(
-            session=session,
-            population_group_descriptions=configuration[
-                "population_group_descriptions"
-            ],
-        )
         self.org = Org(
             session=session,
             datasetinfo=configuration["org"],
-            today=today,
         )
         self.org_type = OrgType(
             session=session,
@@ -87,27 +68,11 @@ class Pipelines:
             datasetinfo=configuration["sector"],
             sector_map=configuration["sector_map"],
         )
-        self.ipc_phase = IpcPhase(
-            session=session,
-            ipc_phase_names=configuration["ipc_phase_names"],
-            ipc_phase_descriptions=configuration["ipc_phase_descriptions"],
-        )
-        self.ipc_type = IpcType(
-            session=session,
-            ipc_type_descriptions=configuration["ipc_type_descriptions"],
-        )
-        self.gender = Gender(
-            session=session,
-            gender_descriptions=configuration["gender_descriptions"],
-        )
-        self.age_range = AgeRange(
-            session=session, age_range_codes=configuration["age_range_codes"]
-        )
 
         Sources.set_default_source_date_format("%Y-%m-%d")
         self.runner = Runner(
             countries,
-            today,
+            today=today,
             errors_on_exit=errors_on_exit,
             scrapers_to_run=scrapers_to_run,
         )
@@ -196,15 +161,9 @@ class Pipelines:
         self.locations.populate()
         self.admins.populate()
         self.metadata.populate()
-        self.population_status.populate()
-        self.population_group.populate()
         self.org.populate()
         self.org_type.populate()
         self.sector.populate()
-        self.ipc_phase.populate()
-        self.ipc_type.populate()
-        self.gender.populate()
-        self.age_range.populate()
 
         if not self.themes_to_run or "population" in self.themes_to_run:
             results = self.runner.get_hapi_results(
@@ -215,8 +174,6 @@ class Pipelines:
                 session=self.session,
                 metadata=self.metadata,
                 admins=self.admins,
-                gender=self.gender,
-                age_range=self.age_range,
                 results=results,
             )
             population.populate()
@@ -249,8 +206,6 @@ class Pipelines:
                 session=self.session,
                 metadata=self.metadata,
                 admins=self.admins,
-                ipc_phase=self.ipc_phase,
-                ipc_type=self.ipc_type,
                 results=results,
             )
             food_security.populate()
@@ -267,11 +222,7 @@ class Pipelines:
                 session=self.session,
                 metadata=self.metadata,
                 admins=self.admins,
-                population_status=self.population_status,
-                population_group=self.population_group,
                 sector=self.sector,
-                gender=self.gender,
-                age_range=self.age_range,
                 results=results,
             )
             humanitarian_needs.populate()
@@ -284,7 +235,7 @@ class Pipelines:
             national_risk = NationalRisk(
                 session=self.session,
                 metadata=self.metadata,
-                admins=self.admins,
+                locations=self.locations,
                 results=results,
             )
             national_risk.populate()