+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.9.0] - 2024-05-14
+
+### Changed
+
+- Many changes to align with V1 of the schema
+
+### Removed
+
+- ipc_phase, ipc_type, age_range, gender classes
+
 ## [0.8.0] - 2024-05-06
 
 ### Changed